+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 - Added extract_tensorboard script.
+- Added plot_logs script.
 
 ## [0.7.3] - 2023-06-03